pyproject.toml
[tool.poetry]
name = "inaturalist"
version = "0.1.3"
description = "iNaturalist Photo Scraper"
authors = ["Alyetama <malyetama@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
requests = ">=2.28.1"
loguru = ">=0.6.0"
minio = ">=7.1.10"
python-dotenv = ">=0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
inat = 'inaturalist:cli.main'
inaturalist = 'inaturalist:cli.main'
